* master:
  simplified Xcode6-Beta4 release note
  @bmunkholm comment updates
  Updated README.md with license and pod tags
  clarified when CGFloat is a float vs a double
  Make message clearer
  - clarified CHANGELOG entry - removed “Note on float” - moved nsnumber_can_be_float() check to nsnumber_is_like_float()
  convert RLM_ARRAY_TYPE documentation to appledoc format
  - extended documentation for RLMConstants, RLMObjectSchema and RLMSchema - tweaked build-docs.sh to reflect these docs changes
  - better document RLMProperty - appledoc no longer ignores RLMProperty - fix a few doc-related issues
  - use CocoaDocs' Dash/DocsForXcode links - no longer build our own docset
  Ignore RLMRealm_Dynamic in appledoc
  updated RealmSwiftTableViewExample to Xcode6-Beta4
  Fixed issue in Swift tests for Xcode6-Beta4: isEqualTo should be isEqual
  typo
  updated CHANGELOG
  fixed issues with Xcode6-Beta4 tests
  WIP support for Xcode6-Beta4
  Allow to build with whatever version of core is in the core folder
  Adding unit tests and a short notice in the documentation about float and double values.
  NSPredicate does not support float values in query strings; only double is. If a value can be casted to a float, and a class has a float property, we can help the user by casting for her but it might lead to loss of precision.

+++ build.sh
@@ -71,7 +71,7 @@ xcode5() {
 }
 
 xcode6() {
-    ln -s /Applications/Xcode6-Beta3.app/Contents/Developer/usr/bin build/bin || exit 1
+    ln -s /Applications/Xcode6-Beta4.app/Contents/Developer/usr/bin build/bin || exit 1
     PATH=./build/bin:$PATH xcodebuild -IDECustomDerivedDataLocation=build/DerivedData $@
 }
 
@@ -173,6 +173,10 @@ case "$COMMAND" in
     # Download Core Library
     ######################################
     "download-core")
+        if [ "$REALM_CORE_VERSION" = "current" ]; then
+            echo "Using version of core already in core/ directory"
+            exit 0
+        fi
         if ! [ -L core ]; then
             echo "core is not a symlink. Deleting..."
             rm -rf core
@@ -181,7 +185,6 @@ case "$COMMAND" in
             download_core
         else
             echo "The core library seems to be up to date."
-            echo "To force an update remove the folder 'core' and rerun."
         fi
         exit 0
         ;;
